Auto-Routing-Bot/requirements.txt
# requirements.txt
flask
python-dotenv
slack_sdk